% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peerGroupComp.r
\name{peerGroupZoo}
\alias{peerGroupZoo}
\title{Zoo Tickers}
\usage{
peerGroupZoo(tickers)
}
\value{
zoo
}
\description{
return a zoo object of time series performance for each
ticker in a vector
}
\examples{
peerGroupZoo(c("IBM", "AAPL", "MSFT"))
}

